requirements.txt
# Core ML and Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
torch==2.0.1
torchvision==0.15.2

# Snowflake and Database
snowflake-connector-python==3.2.1
snowflake-ml-python==1.0.12
snowflake-snowpark-python==1.8.0

# Data Generation and Utilities
faker==19.3.1
python-dateutil==2.8.2
pytz==2023.3

# Visualization and Analysis
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Configuration and Logging
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.0

# Streaming and API
requests==2.31.0
fastapi==0.100.1
uvicorn==0.23.2
